[Misc] Move utils to avoid conflicts with stdlib
