+++ vllm/utils/collection_utils.py
